+++ packages/db/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @lunora/db [1.0.0-alpha.93](https://github.com/anolilab/lunora/compare/@lunora/db@1.0.0-alpha.92...@lunora/db@1.0.0-alpha.93) (2026-09-08)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.90
+
+## @lunora/db [1.0.0-alpha.92](https://github.com/anolilab/lunora/compare/@lunora/db@1.0.0-alpha.91...@lunora/db@1.0.0-alpha.92) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.89
+
 ## @lunora/db [1.0.0-alpha.91](https://github.com/anolilab/lunora/compare/@lunora/db@1.0.0-alpha.90...@lunora/db@1.0.0-alpha.91) (2026-09-07)