+++ packages/react/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/react [1.0.0-alpha.95](https://github.com/anolilab/lunora/compare/@lunora/react@1.0.0-alpha.94...@lunora/react@1.0.0-alpha.95) (2026-09-08)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.90
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.46
+
+## @lunora/react [1.0.0-alpha.94](https://github.com/anolilab/lunora/compare/@lunora/react@1.0.0-alpha.93...@lunora/react@1.0.0-alpha.94) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.89
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.45
+
 ## @lunora/react [1.0.0-alpha.93](https://github.com/anolilab/lunora/compare/@lunora/react@1.0.0-alpha.92...@lunora/react@1.0.0-alpha.93) (2026-09-07)