+++ packages/replica/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @lunora/replica [1.0.0-alpha.89](https://github.com/anolilab/lunora/compare/@lunora/replica@1.0.0-alpha.88...@lunora/replica@1.0.0-alpha.89) (2026-09-08)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.110
+
+## @lunora/replica [1.0.0-alpha.88](https://github.com/anolilab/lunora/compare/@lunora/replica@1.0.0-alpha.87...@lunora/replica@1.0.0-alpha.88) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.109
+
 ## @lunora/replica [1.0.0-alpha.87](https://github.com/anolilab/lunora/compare/@lunora/replica@1.0.0-alpha.86...@lunora/replica@1.0.0-alpha.87) (2026-09-07)