+++ packages/scheduler/CHANGELOG.md
@@ -1,3 +1,7 @@
+## @lunora/scheduler [1.0.0-alpha.60](https://github.com/anolilab/lunora/compare/@lunora/scheduler@1.0.0-alpha.59...@lunora/scheduler@1.0.0-alpha.60) (2026-09-08)
+
+## @lunora/scheduler [1.0.0-alpha.59](https://github.com/anolilab/lunora/compare/@lunora/scheduler@1.0.0-alpha.58...@lunora/scheduler@1.0.0-alpha.59) (2026-09-07)
+
 ## @lunora/scheduler [1.0.0-alpha.58](https://github.com/anolilab/lunora/compare/@lunora/scheduler@1.0.0-alpha.57...@lunora/scheduler@1.0.0-alpha.58) (2026-09-07)