+++ packages/vue/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/vue [1.0.0-alpha.126](https://github.com/anolilab/lunora/compare/@lunora/vue@1.0.0-alpha.125...@lunora/vue@1.0.0-alpha.126) (2026-09-08)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.90
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.46
+
+## @lunora/vue [1.0.0-alpha.125](https://github.com/anolilab/lunora/compare/@lunora/vue@1.0.0-alpha.124...@lunora/vue@1.0.0-alpha.125) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.89
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.45
+
 ## @lunora/vue [1.0.0-alpha.124](https://github.com/anolilab/lunora/compare/@lunora/vue@1.0.0-alpha.123...@lunora/vue@1.0.0-alpha.124) (2026-09-07)